sample-app/requirements.txt
avro-python3==1.8.2
kafka-python==1.4.6
requests 
structlog 
colorama
pydash